+++ tests/testthat/test-fit_mpm.R
@@ -1,24 +1,23 @@
 context("test fit_mpm")
 
-## fit_mpm output is a compound tibble with class `fG_mpm`
+## fit_mpm output is a compound tibble with class `mpm`
 ##  tests expect that fit$mpm are 8-element lists (if optimiser does not crash)
 ##  that have S3 class foieGras
 
-## generate fG_ssm obj a quickly as possible
-## have to do this to avoid error when calling st_transform on platforms running older GDAL versions (sese2 is highly sub-sampled for this purpose)
-xs <- fit_ssm(sese2, spdf=FALSE, model = "crw", time.step=72, 
-              control = ssm_control(verbose = 0))
-dmp <- grab(xs, "predicted", as_sf = FALSE)
-dmp <- dmp[, c("id", "date", "lon", "lat")]
-
-test_that("fit_mpm returns fG_mpm list w 8 elements", {
+## generate ssm obj as quickly as possible
+test_that("fit_mpm returns mpm list w 8 elements", {
+  xs <- fit_ssm(sese2, spdf=FALSE, model = "crw", time.step=72, 
+                control = ssm_control(verbose = 0))
+  dmp <- grab(xs, "predicted")
   fmp <- fit_mpm(dmp, model = "jmpm", control = mpm_control(verbose = 0))
-  expect_s3_class(fmp, "fG_mpm")
+  expect_s3_class(fmp, "mpm_df")
   expect_equal(length(fmp$mpm), 1)
 })
 
 
-test_that("fit_mpm fits directly to an fG_ssm object", {
+test_that("fit_mpm fits directly to an ssm object", {
+  xs <- fit_ssm(sese2, spdf=FALSE, model = "crw", time.step=72, 
+                control = ssm_control(verbose = 0))
   fmp <- fit_mpm(xs, model = "jmpm", control = 
                    mpm_control(verbose = 0, 
                                optim = "optim", 
@@ -26,6 +25,6 @@ test_that("fit_mpm fits directly to an fG_ssm object", {
                                lower = list(l_sigma_g = -1)
                                )
                  )
-  expect_s3_class(fmp, "fG_mpm")
+  expect_s3_class(fmp, "mpm_df")
   expect_equal(length(fmp$mpm), 1)
 })