+++ tests/testthat/test-fit_ssm.R
@@ -42,7 +42,7 @@ test_that("fit_ssm defaults + crw + KF return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, model = "crw", time.step = 72, control = ssm_control(optim = "nlminb", verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 
@@ -51,7 +51,7 @@ test_that("fit_ssm defaults + crw + LS return foieGras list w 15 elements", {
   f <- fit_ssm(dls, model = "crw", time.step = 72, control = ssm_control(verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## minimum specified arguments - rw + KF
@@ -60,7 +60,7 @@ test_that("fit_ssm defaults + rw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, model = "rw", time.step = 72, map = list(psi = as.factor(NA)), control = ssm_control(verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## minimum specified arguments - rw + LS
@@ -69,7 +69,7 @@ test_that("fit_ssm defaults + rw return foieGras list w 15 elements", {
   f <- fit_ssm(dls, model = "rw", time.step = 72, map = list(psi = as.factor(NA)), control = ssm_control(verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## low vmax - crw
@@ -78,7 +78,7 @@ test_that("fit_ssm vmax + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, vmax=5, model = "crw", time.step = 72, control = ssm_control(optim = "nlminb", verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## low vmax, ang=c(15,25) - crw
@@ -87,7 +87,7 @@ test_that("fit_ssm vmax,ang + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, vmax=5, ang=c(15,25), model = "crw", time.step = 72, control = ssm_control(optim = "nlminb", verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## low vmax, ang=c(15,25), min.dt=120 - crw
@@ -96,7 +96,7 @@ test_that("fit_ssm vmax,ang,min.dt + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, vmax=5, ang=c(15,25), min.dt=120, model = "crw", time.step = 72, control = ssm_control(optim = "nlminb", verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## turn speed filter off - crw
@@ -105,7 +105,7 @@ test_that("fit_ssm no spd filter + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, spdf=FALSE, model = "crw", time.step = 72, control = ssm_control(optim = "nlminb", verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## step through some SSM arguments
@@ -115,7 +115,7 @@ test_that("fit_ssm optim + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, model = "crw", time.step = 72, control = ssm_control(verbose = 1, optim="nlminb"))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## fit to full data - rw
@@ -124,7 +124,7 @@ test_that("fit_ssm full data + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, fit.to.subset=FALSE, model = "rw", time.step = 72, control = ssm_control(verbose = 0))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## turn trace on - crw
@@ -133,7 +133,7 @@ test_that("fit_ssm verbose + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, model = "crw", time.step = 72, control = ssm_control(optim = "nlminb", verbose=2))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 # fiddle w inner control list - crw
@@ -142,7 +142,7 @@ test_that("fit_ssm inner.control + crw return foieGras list w 15 elements", {
   f <- fit_ssm(dkf, inner.control=list(maxit=200), model = "crw", time.step = 72, control = ssm_control(verbose = 0, optim = "nlminb"))
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })
 
 ## specify lower, upper parameter bounds
@@ -156,5 +156,5 @@ test_that("fit_ssm multi-track returns foieGras lists w 15 elements", {
   expect_equal(nrow(f), 1)
   expect_s3_class(f$ssm[[1]], "ssm")
   expect_equal(length(f$ssm[[1]]), 15)
-  expect_s3_class(f, "fG_ssm")
+  expect_s3_class(f, "ssm_df")
 })