+++ tests/testthat/test-plot_mpm.R
@@ -1,14 +1,14 @@
 context("test plot_mpm")
 
 ## expect plot is silent
-data(xm)
-## generate fG_ssm obj a quickly as possible
+## generate ssm obj a quickly as possible
 ## have to do this to avoid error when calling st_transform on platforms running older GDAL versions (sese2 is highly sub-sampled for this purpose)
 xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, 
-              control = ssm_control(se = FALSE, verbose = 0))
+              control = ssm_control(verbose = 0))
+xm <- fit_mpm(xs, model = "jmpm")
 
 ## plot gamma time-series on a single page
-tp <- plot(xm, pages = 1, ncol = 1)
+tp <- plot(xm, pages = 1)
 test_that("plot completes silently - fitted", {
   expect_s3_class(tp, c("gg","ggplot"))
 })