+++ tests/testthat/test-plot_ssm.R
@@ -1,6 +1,6 @@
 context("test plot_ssm")
 
-## generate fG_ssm obj a quickly as possible
+## generate ssm obj a quickly as possible
 ## have to do this to avoid error when calling st_transform on platforms running older GDAL versions (sese2 is highly sub-sampled for this purpose)
 xs <- fit_ssm(sese2, spdf=FALSE, model = "rw", time.step=72, 
               control = ssm_control(se = FALSE, verbose = 0))